Force Tip Type to Unknown
Added line to start() that forces _tip_type to TipType::UNKNOWN with the rest of the variable initialization. This forces the iron to limit max output power to 150mW until a tip type is determined (RTM or RTU), which prevents 150W being sent to a 40W tip for even a single cycle when tips are changed.